Rename: 전체적인 폴더구조 및 폴더명 변경
