+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { useDeviceStore } from '@/store/useDeviceStore';
+import { useDeviceStore } from '@/stores/useDeviceStore';
 import { useEffect } from 'react';
 import AppRouter from './AppRouter';
