+++ src/components/Header/index.ts
@@ -1,2 +1,2 @@
 export { default as Header } from './Header';
-export * from './Mobile';
+export * from './mobile';
