+++ src/hooks/queries/usePostsQuery.ts
@@ -1,4 +1,4 @@
-import { getPosts } from '@/api/home/getPosts';
+import { getPosts } from '@/api/post/getPosts';
 import { QUERY_KEYS } from '@/constants/api';
 import { useQuery } from '@tanstack/react-query';
