+++ src/pages/CreatePostPage.tsx
@@ -1,13 +1,13 @@
 import { Button } from '@/components/common/Button';
 import Input from '@/components/common/Input';
 import TextArea from '@/components/common/TextArea';
-import MobileEditorHeader from '@/components/Header/Mobile/MobileEditorHeader';
-import { usePostDetailQuery } from '@/hooks/queries/usePostDetailQuery';
-import { useDeviceStore } from '@/store/useDeviceStore';
-import { useEffect, useState, useMemo } from 'react';
-import { useSearchParams } from 'react-router-dom';
+import MobileEditorHeader from '@/components/header/mobile/MobileEditorHeader';
 import { useCreatePostMutation } from '@/hooks/queries/useCreatePostMutation';
 import { useEditPostMutation } from '@/hooks/queries/useEditPostMutation';
+import { usePostDetailQuery } from '@/hooks/queries/usePostDetailQuery';
+import { useDeviceStore } from '@/stores/useDeviceStore';
+import { useEffect, useMemo, useState } from 'react';
+import { useSearchParams } from 'react-router-dom';
 
 interface FormErrors {
   title?: string;
