+++ src/stores/useDeviceStore.ts
